src/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
dateparser==0.7.1
idna==2.7
more-itertools==4.3.0
phonetic-alphabet==0.1.0
pluggy==0.7.1
py==1.6.0
pytest==3.8.1
python-dateutil==2.7.5
pytz==2018.7
regex==2018.11.22
requests==2.20.0
ruamel.yaml==0.15.78
six==1.11.0
timeago==1.0.8
tzlocal==1.5.1
urllib3==1.23
pycodestyle==2.5.0
